backend/requirements.txt
Flask
flask_cors
python-dotenv
streamlit
google-generativeai==0.3.2
requests
cloudinary
PyMuPDF
Werkzeug